requirements.txt
anyio==4.2.0
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
feedparser==6.0.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
oauthlib==3.2.2
PyJWT==2.8.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
sgmllib3k==1.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.4.4
tweepy==4.14.0
typing_extensions==4.9.0
urllib3==2.2.0
